# HPCC-17914 Optimize the performance for WildMatch