@ghalliday 2017-07-05
@richardkchapman please review.  Also, can you think of any other test cases that I haven't included.  Some sample timing comparisons for the timing results:

Before:
10000: 53 ms
100000: 2389 ms
1000000: 227851 ms

After:
10000: 7 ms
100000: 380 ms
1000000: 26029 ms