@jakesmith 2017-07-11
@ghalliday - can't see any issues, test coverage looks good to me.
A couple of trivial comments.